Impl changes for AuthenticationInterface
