+++ crates/diesel_models/src/query/authentication.rs
@@ -2,54 +2,11 @@ use diesel::{associations::HasTable, BoolExpressionMethods, ExpressionMethods};
 
 use super::generics;
 use crate::{
-    authentication::{
-        Authentication, AuthenticationNew, AuthenticationUpdate, AuthenticationUpdateInternal,
-    },
+    authentication::{Authentication, AuthenticationNew, AuthenticationUpdateInternal},
     errors,
     schema::authentication::dsl,
     PgPooledConn, StorageResult,
 };
-
-impl Authentication {
-    pub async fn update_by_merchant_id_authentication_id_internal(
-        conn: &PgPooledConn,
-        merchant_id: &common_utils::id_type::MerchantId,
-        authentication_id: &common_utils::id_type::AuthenticationId,
-        update: AuthenticationUpdateInternal,
-    ) -> StorageResult<Self> {
-        match generics::generic_update_with_unique_predicate_get_result::<
-            <Self as HasTable>::Table,
-            _,
-            _,
-            _,
-        >(
-            conn,
-            dsl::merchant_id
-                .eq(merchant_id.to_owned())
-                .and(dsl::authentication_id.eq(authentication_id.to_owned())),
-            update.clone(),
-        )
-        .await
-        {
-            Err(error) => match error.current_context() {
-                errors::DatabaseError::NotFound => Err(error.attach_printable(
-                    "Authentication with the given Authentication ID does not exist",
-                )),
-                errors::DatabaseError::NoFieldsToUpdate => {
-                    generics::generic_find_one::<<Self as HasTable>::Table, _, _>(
-                        conn,
-                        dsl::merchant_id
-                            .eq(merchant_id.to_owned())
-                            .and(dsl::authentication_id.eq(authentication_id.to_owned())),
-                    )
-                    .await
-                }
-                _ => Err(error),
-            },
-            result => result,
-        }
-    }
-}
 
 impl AuthenticationNew {
     pub async fn insert(self, conn: &PgPooledConn) -> StorageResult<Authentication> {
@@ -60,9 +17,9 @@ impl AuthenticationNew {
 impl Authentication {
     pub async fn update_by_merchant_id_authentication_id(
         conn: &PgPooledConn,
-        merchant_id: common_utils::id_type::MerchantId,
-        authentication_id: common_utils::id_type::AuthenticationId,
-        authorization_update: AuthenticationUpdate,
+        merchant_id: &common_utils::id_type::MerchantId,
+        authentication_id: &common_utils::id_type::AuthenticationId,
+        authentication_update: AuthenticationUpdateInternal,
     ) -> StorageResult<Self> {
         match generics::generic_update_with_unique_predicate_get_result::<
             <Self as HasTable>::Table,
@@ -74,7 +31,7 @@ impl Authentication {
             dsl::merchant_id
                 .eq(merchant_id.to_owned())
                 .and(dsl::authentication_id.eq(authentication_id.to_owned())),
-            AuthenticationUpdateInternal::from(authorization_update),
+            authentication_update,
         )
         .await
         {
